feat: update chatbot
