+++ src/services/axios/actions/user.action.ts
@@ -1,5 +1,6 @@
 import { client } from "@/services/axios"
 import { ApiResponse, User } from "@/types/api/auth"
+import { ChatRequest, ChatResponse } from "@/types/api/chatbot"
 import userEndpoint from "../endpoints/user.endpoint"
 
 class UserApi {
@@ -12,14 +13,14 @@ class UserApi {
         }
     }
 
-    // async chatBot(payload) {
-    //   try {
-    //     const res = await client.post(userEndpoint.chatbot, payload)
-    //     return res
-    //   } catch (error) {
-    //     console.log(error)
-    //   }
-    // }
+    async chatBot(payload: ChatRequest): Promise<ChatResponse> {
+        try {
+            const res = await client.post(userEndpoint.chatbot, payload)
+            return res.data
+        } catch (error) {
+            throw error
+        }
+    }
 }
 
 export default new UserApi()
